+++ planning/autoware_tensorrt_vad/src/vad_node.cpp
@@ -21,7 +21,7 @@
 
 namespace autoware::tensorrt_vad
 {
-std::pair<Eigen::Matrix4f, Eigen::Matrix4f> get_transform_matrix(
+std::pair<Eigen::Matrix4d, Eigen::Matrix4d> get_transform_matrix(
   const nav_msgs::msg::Odometry & msg)
 {
   // Extract position
@@ -30,31 +30,31 @@ std::pair<Eigen::Matrix4f, Eigen::Matrix4f> get_transform_matrix(
   double z = msg.pose.pose.position.z;
 
   // Create Eigen quaternion and normalize it
-  Eigen::Quaternionf q = std::invoke([&msg]() -> Eigen::Quaternionf {
+  Eigen::Quaterniond q = std::invoke([&msg]() -> Eigen::Quaterniond {
     double qx = msg.pose.pose.orientation.x;
     double qy = msg.pose.pose.orientation.y;
     double qz = msg.pose.pose.orientation.z;
     double qw = msg.pose.pose.orientation.w;
 
     // Create Eigen quaternion and normalize it
-    Eigen::Quaternionf q(qw, qx, qy, qz);
-    return (q.norm() < std::numeric_limits<float>::epsilon()) ? Eigen::Quaternionf::Identity()
-                                                              : q.normalized();
+    Eigen::Quaterniond q(qw, qx, qy, qz);
+    return (q.norm() < std::numeric_limits<double>::epsilon()) ? Eigen::Quaterniond::Identity()
+                                                               : q.normalized();
   });
 
   // Rotation matrix (3x3)
-  Eigen::Matrix3f R = q.toRotationMatrix();
+  Eigen::Matrix3d R = q.toRotationMatrix();
 
   // Translation vector
-  Eigen::Vector3f t(x, y, z);
+  Eigen::Vector3d t(x, y, z);
 
   // Base_link → Map (forward)
-  Eigen::Matrix4f bl2map = Eigen::Matrix4f::Identity();
+  Eigen::Matrix4d bl2map = Eigen::Matrix4d::Identity();
   bl2map.block<3, 3>(0, 0) = R;
   bl2map.block<3, 1>(0, 3) = t;
 
   // Map → Base_link (inverse)
-  Eigen::Matrix4f map2bl = Eigen::Matrix4f::Identity();
+  Eigen::Matrix4d map2bl = Eigen::Matrix4d::Identity();
   map2bl.block<3, 3>(0, 0) = R.transpose();
   map2bl.block<3, 1>(0, 3) = -R.transpose() * t;
